xads-scraper/requirements.txt
google-genai>=0.8.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dateutil>=2.8.2
